Optimize gulpfile.js

+++ gulpfile.js
@@ -219,20 +219,22 @@ gulp.task('5-copy', () => copyFiles());
 gulp.task('5-copy-dep', gulp.series('3-build-dep', '5-copy'));
 
 gulp.task('6-patch', done => {
-    let code = fs.readFileSync(__dirname + '/admin/tab.html').toString('utf8');
-    code = code.replace(/<script>var head=document\.getElementsByTagName\("head"\)\[0\][^<]+<\/script>/,
-        `<script type="text/javascript" src="./lib/js/socket.io.js"></script>`);
-    // add monaco script at the end
-    if (!code.includes(`<script type="text/javascript" src="vs/loader.js"></script><script type="text/javascript" src="vs/configure.js"></script>`)) {
-        code = code.replace('</body></html>', `<script type="text/javascript" src="vs/loader.js"></script><script type="text/javascript" src="vs/configure.js"></script></body></html>`);
-    }
+    if (fs.existsSync(__dirname + '/admin/tab.html')) {
+        let code = fs.readFileSync(__dirname + '/admin/tab.html').toString('utf8');
+        code = code.replace(/<script>var head=document\.getElementsByTagName\("head"\)\[0\][^<]+<\/script>/,
+            `<script type="text/javascript" src="./lib/js/socket.io.js"></script>`);
+        // add monaco script at the end
+        if (!code.includes(`<script type="text/javascript" src="vs/loader.js"></script><script type="text/javascript" src="vs/configure.js"></script>`)) {
+            code = code.replace('</body></html>', `<script type="text/javascript" src="vs/loader.js"></script><script type="text/javascript" src="vs/configure.js"></script></body></html>`);
+        }
 
-    fs.writeFileSync(__dirname + '/admin/tab.html', code);
+        fs.writeFileSync(__dirname + '/admin/tab.html', code);
+    }
 
     done();
 });
 
-gulp.task('6-patch-dep',  gulp.series('5-copy', '6-patch'));
+gulp.task('6-patch-dep',  gulp.series('5-copy-dep', '6-patch'));
 
 gulp.task('webserver', () => {
     connect.server({
